+++ src/cgame/cg_players.cpp
@@ -2172,7 +2172,7 @@ static void CG_PlayerWWSmoothing( centity_t *cent, vec3_t in[ 3 ], vec3_t out[ 3
 			           DotProduct( inAxis[ 1 ], lastAxis[ 1 ] ) +
 			           DotProduct( inAxis[ 2 ], lastAxis[ 2 ] );
 
-			rotAngle = RAD2DEG( acos( ( rotAngle - 1.0f ) / 2.0f ) );
+			rotAngle = RAD2DEG( acosf( ( rotAngle - 1.0f ) / 2.0f ) );
 
 			CrossProduct( lastAxis[ 0 ], inAxis[ 0 ], temp );
 			VectorCopy( temp, rotAxis );
@@ -2206,7 +2206,7 @@ static void CG_PlayerWWSmoothing( centity_t *cent, vec3_t in[ 3 ], vec3_t out[ 3
 		if ( cg.time < cent->pe.sList[ i ].time + MODEL_WWSMOOTHTIME )
 		{
 			stLocal = 1.0f - ( ( ( cent->pe.sList[ i ].time + MODEL_WWSMOOTHTIME ) - cg.time ) / MODEL_WWSMOOTHTIME );
-			sFraction = - ( cos( stLocal * M_PI ) + 1.0f ) / 2.0f;
+			sFraction = - ( cosf( stLocal * M_PI ) + 1.0f ) / 2.0f;
 
 			RotatePointAroundVector( outAxis[ 0 ], cent->pe.sList[ i ].rotAxis,
 			                         inAxis[ 0 ], sFraction * cent->pe.sList[ i ].rotAngle );
